backend/requirements.txt
# Core Django
Django==5.0.6

# REST Framework
djangorestframework==3.15.2
django-filter==24.3

# Authentication
djangorestframework-simplejwt==5.3.1
djoser==2.3.1

# Database
psycopg2-binary==2.9.9

# Media & File Handling
Pillow==10.3.0
django-storages==1.14.3  # if using AWS S3/Cloud Storage

# Environment Variables
python-decouple==3.8

# CORS
django-cors-headers==4.3.1

# Static Files in Production
whitenoise==6.7.0

# WSGI Server for Production
gunicorn==22.0.0

# Optional Utilities
drf-yasg==1.21.7  # API documentation (Swagger / ReDoc)

cloudinary

django-cloudinary-storage

django-environ
